move userpreferenc test in UserPreferencesSerialisationTest
